[ADD] charts: refactored the css
