--- fitness_app/src/components/charts/UserAnotherChart.module.css
@@ -1,19 +1,0 @@
-.userAnotherChartContainer {
-    display: flex;
-    height: 35vh;
-    width: 80vh;
-    background-color:var(--white-color);
-    box-shadow: 0 0 6px 2px rgba(0,0,0,.1);
-    border-bottom: 2px solid transparent;
-    border-radius: 10px;
-}
-
-@media screen and (max-width: 768px) {
-    .userPerformanceChartContainer {
-        justify-content: center;
-        align-items: center;
-        width: fit-content;
-    }
-
-}
-
