--- fitness_app/src/components/charts/UserCaloriesChart.module.css
@@ -1,12 +1,0 @@
-.userCaloriesChart {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 300px; /* Add 'px' to specify height in pixels */
-    width: 100vh; /* Use a relative width to fit the available space */
-    max-width: 300px; /* Add a maximum width to prevent items from getting too wide */
-    background-color: var(--white-color);
-    box-shadow: 0 0 6px 2px rgba(0,0,0,.1);
-    border-bottom: 2px solid transparent;
-    border-radius: 10px;
-  }
