+++ fitness_app/src/components/charts/UserPerformanceChart.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useRef } from "react";
-import { Chart } from "chart.js";
-import styles from "./UserPerformanceChart.module.css";
+import { Chart } from "chart.js/auto";
+import styles from "./charts.module.css";
 import { useSession } from "../../supabase/sessionContext";
 import { supabase } from "../../supabase/client";
