+++ fitness_app/src/components/charts/charts.module.css
@@ -1,3 +1,13 @@
+.userAnotherChartContainer {
+    display: flex;
+    height: 35vh;
+    width: 80vh;
+    background-color:var(--white-color);
+    box-shadow: 0 0 6px 2px rgba(0,0,0,.1);
+    border-bottom: 2px solid transparent;
+    border-radius: 10px;
+}
+
 .userPerformanceChartContainer {
     display: flex;
     height: 35vh;
@@ -12,8 +22,15 @@
     .userPerformanceChartContainer {
         justify-content: center;
         align-items: center;
-        width: fit-content;
+        width: 48vh;
+    }
+    .userAnotherChartContainer {
+        justify-content: center;
+        align-items: center;
+        width: 48vh;
     }
 
 }
+
+
