+++ fitness_app/src/pages/Dashboard/dashboard.module.css
@@ -1,24 +1,14 @@
-/* CSS for the dashboard and containers */
 .dashboard {
     display: flex;
     flex-direction: row;
     overflow: hidden;
     background-color: var(--primary-color);
+    gap: 70px;
   }
-  
-  .bottomChart {
-    display: flex;
-    flex-direction: row;
-  }
-  
-  .sidebarContainer {
-    display: flex;
-    position: relative;
-    height: 100vh;
-  }
-  
+
   .widgetsContainer {
     display: flex;
+    flex: 1;
     flex-direction: column;
     padding-top: 50px;
     padding-left: 30px;
@@ -32,37 +22,42 @@
   .topContainer {
     display: flex;
     flex-direction: row;
-    gap: 200px; /* Reduce the gap for better spacing */
+    gap: 200px; 
     flex-wrap: nowrap;
-    overflow: auto; /* Add horizontal and vertical scrolling for smaller screens if necessary */
+    overflow: auto; 
   }
   
   .bottomContainer {
     display: flex;
-    gap: 100px;
-
+    gap: 20px;
   }
+
+  .bottomChart {
+    display: flex;
+    flex-direction: row;
+  }
+  
   
   @media screen and (max-width: 768px) {
     .dashboard {
-      flex-direction: row; /* Change to column layout for small screens */
+      flex-direction: row; 
     }
   
     .widgetsContainer {
-      padding-top: 20px; /* Adjust padding for better spacing on small screens */
+      padding-top: 20px;
       padding-left: 65px;
     }
   
     .topContainer {
-      flex-direction: column; /* Display items in a column on small screens */
-      gap: 20px; /* Adjust the gap between items as needed */
-      overflow: hidden; /* Turn off both horizontal and vertical scrolling */
+      flex-direction: column; 
+      gap: 20px; 
+      overflow: hidden; 
     }
 
     .bottomContainer {
-      flex-direction: column; /* Display items in a column on small screens */
-      gap: 20px; /* Adjust the gap between items as needed */
-      overflow: hidden; /* Turn off both horizontal and vertical scrolling */
+      flex-direction: column;
+      gap: 20px; 
+      overflow: hidden; 
       height: fit-content;
     }
